Add highest available version to the report view
Highest version includes "~" versions but not ~9999 (generally not
useful).
Relies on eix listing the version numbers in ascending order.
Also list show all installed packages in bold, not just those which are
world packages.
The tests change for the new data column, and "test.py" now does a dump
of the same pb file as used by the tests.